fix(scripts): select the last matching dict key; split fact_detector_misuse.py
_static_subscript_element()'s dict-key resolution used a forward scan,
returning the first matching key rather than the last -- a real Python
dict literal keeps the last value for a repeated (or merely ==-equal,
e.g. 1/True) key. `{"x": Fact.present(1), "x": 0}["x"]` was reported as
a misuse site even though the value actually selected at runtime is 0.

Fixed by scanning the full zip(keys, values) sequence and keeping the
last match rather than returning on the first one -- the existing
==-based key comparison already makes the 1/True collision resolve
correctly too, with no separate logic needed.

Verified against 8 cases via direct AST reproduction before writing
tests: the reported case, the reversed case (Fact value last, correctly
flagged), both directions of the True/1 collision, a triple-duplicate
key, and regression guards.

This fix pushed fact_detector_misuse.py to 1999/2000 lines -- genuinely
out of headroom after two prior rounds flagged this as imminent. Split
proactively per standing instruction, mirroring the exact precedent
fact_detector_misuse_scope.py's own earlier split established: moved
the entire "does this expression/annotation/name resolve to a Fact[T]
value" machinery (FACT_FIELD_NAMES, _imported_fact_aliases,
_is_fact_typed_expr, _static_display_elements,
_static_subscript_element, _admissible_loop_element,
_candidate_resolves_to_fact, _annotation_head_name,
_is_fact_typed_annotation, _fact_aliases) to a new sibling module,
fact_detector_misuse_aliases.py -- mechanical extraction, unchanged
function bodies. _iter_default_subtree (and the
_SCOPE_INTRODUCING_NODE_TYPES constant it uses) turned out to be needed
by both remaining files, so it relocated a second level up into
fact_detector_misuse_scope.py rather than being duplicated or creating
an import cycle. Re-exported names use the explicit `X as X` spelling
checker_policy.py already uses for ChangeKind, so every existing
`from .fact_detector_misuse import FACT_FIELD_NAMES` call site
(including this check's own test suite) is unaffected.

Resulting line counts: fact_detector_misuse.py 381 lines (from 1999),
fact_detector_misuse_aliases.py 1614 lines (new),
fact_detector_misuse_scope.py 1434 lines (from 1369) -- all three with
ample headroom now. scripts/CLAUDE.md's inventory table updated.

New tests: TestLastMatchingKeyWins (6 tests) in
tests/test_fact_detector_misuse_static_subscript.py. Full
fact_detector_misuse test group (364 tests) re-verified green after the
split, plus module-reload and mypy/ruff checks on all three files.
